Changed the Service Execution Method
Rather than using the "apply" method, the method "call" should be used.
This is just a precaution to prevent any future issues with case classes
or something similar.
Furthermore, the Predicate traits were enriched by references and test
methods. In order to test a predicate, a map can be provided that maps
the values to the corresponding reference types.