Rewrote buildingblock concept
After consideration, it becomes obvious that it is not at all necessary
to use typed input and output references for services. Instead, a new
type called "Symbol" is introduced. These symbols are primitive objects
that can be compared via identity (==).
Furthermore, Services, Converters, and Predicates are traits that can be
added to existing implementations. Predicates do not have to be services
(since services require pre and post conditions). Also, Converters don't
have to be Services either, but do also require pre and post conditions.