Moved Datatypes to an Independent package
The repositories are meant to be constructed on a domain specific basis.
Hence, we try to implement everything as domain independent as possible.
This allows for a more generic construction of new repositories by
simply exchanging the domain specific part.
The datatypes are the most important component of the domain dependent
part.