Changing the Service Interface
The old service interface was not sufficient enough. Rather than just
having a simple service trait, it is neccessary to have both a common
interface for external services (ExternalService) and for components
that hold a reference to these services (ExternalServiceReference).
The most important such components are the service entries themselves.
Each external service has to be published to the repository by creating
a new service entry (ServiceEntry). This entry contains all metadata
required to make it repository compliant.