Rework of the Predicates Traits
This is a major change on the way how predicates work. Firstmost, there
are two different type of predicates: unary and binary predicates. These
different types are important because they work very differently. We
must later be able to compare the input of different predicates. If some
take tuples this will become very hard. Also, having these two different
predicates does not hinder us from staying with unary services.
There are three major predicates "And", "Or", and "Not". These
predicates can be used to combine other predicates. They are not
implemented using some service and hence have to be treated independent.
For predicates that are implemented using services two corresponding
classes exist "PredicateService1" and "PredicateService2". These classes
are wrapped in entry classes "PredicateServiceEntry1" and
"PredicateServiceEntry2" that are bound to some specific implementation
and scenario.
Simple "PredicateEntry1" and "PredicateEntry2" classes exist for
allowing the onging construction of new predicate service entries when
encountering new scenarios (having to describe other services).

At a later point, it will be necessary to match these different types of
predicates in order to create predicatesymbols. Hence, they are created
using case classes.