Starting the Repository
This commit holds the basic implementation of the repository containing
all entries (predicates, services, converters). For now, the predicates
are hardcoded properties. This might have to change in the future, but
is sufficient for now.
The Symbol needs to be changed to allow for covariant type parameter.
Otherwise, the "Equals" predicate would only facilitate one single type
of object equality (e.g. Flights) rather than any.