Implementation of Predicatesymbols
Predicatesymbols are being constructed on the fly using a Generator.
This generator checks the identity of the predicate parameters and thus
identifies the input/output reference in the replacement graph.
The position of the reference becomes part of the predicatesymbol. By
doing that, the system provides means for symbolized consideration of
predicate parameters.